Move Cwp methods into new file
